third_party/WebKit/LayoutTests/platform/mac/fast/events/touch/gesture/touch-gesture-noscroll-body-yhidden-expected.txt

This tests that a page cannot be scrolled vertically with touch (but can still be scrolled horizontally) if its body has style overflow-y:hidden. The scroll events in this test have both an x and y component.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


first gesture
FAIL document.body.scrollTop should be 0. Was 110.
FAIL document.body.scrollLeft should be 55. Was 0.
PASS wheelEventsOccurred is 0
second gesture
FAIL document.body.scrollTop should be 0. Was 205.
FAIL document.body.scrollLeft should be 97. Was 0.
PASS wheelEventsOccurred is 0
scroll event 0+> [object HTMLDocument]
PASS scrollEventsOccurred is 1
PASS successfullyParsed is true

TEST COMPLETE
